Value At Risk notebook potentially ready.
The major change is surely the creation of the value_at_risk notebook,
in my opinion possibly ready to be read from an external audience.
The mathematical details have been reduced, the exposition made more
linear and at the end there is the inclusion of a simple example
working on the DAX index.
Finally, the .gitignore file has been updated in order to *exclude*
jupyter checkpoints.

    biagio